refactor: moved session locks to cli mod
