+++ crates/baro-tui/src/cli/mod.rs
@@ -1,1 +1,2 @@
 pub mod cli;
+pub mod session;
